ai-backend/requirements.txt
torch
torchvision
numpy
opencv-python
Pillow
matplotlib
seaborn
jupyter
notebook
ipykernel
pandas
scikit-learn
tqdm
python-dotenv
requests
aiohttp
websockets
pytest
pytest-asyncio
pytest-cov
black
flake8
mypy
datasets
opencv-python
albumentations
ultralytics
wandb
imagehash

fastapi
uvicorn[standard]
pillow
numpy
opencv-python-headless
ultralytics
torch
python-multipart
